Add extra_float_digits connection option defaulted to 3
Add a connection option for setting extra_float_digits and set its
default value to 3. This allows for more accurate representation of
large floating point values.

The option is sent in the connection startup packet so as not to incur
an extra round trip. This will not work with older database that do
not support setting the extra_float_digits connection parameter in the
startup packet (PostgreSQL versions <=8.4).

Closes #730

WARNING: This is a breaking change and will prevent connecting to older
         (<=8.4) databases unless extra_float_digits is set to 0.